+++ packages/server/utils/getReqAuth.ts
@@ -1,5 +1,4 @@
 import type {HttpRequest} from 'uWebSockets.js'
-import {PAT_PREFIX} from '../graphql/public/applyScopeDirective'
 import {getAuthTokenFromCookie} from './authCookie'
 import getVerifiedAuthToken from './getVerifiedAuthToken'
 
@@ -7,7 +6,7 @@ const getReqAuth = (req: HttpRequest, logErrors = true) => {
   // mattermost plugin cannot use the `authorization` header directly
   const authHeader = req.getHeader('x-application-authorization') || req.getHeader('authorization')
   const headerToken = authHeader.slice(7)
-  if (headerToken?.startsWith(PAT_PREFIX)) return null
+
   const cookieToken = getAuthTokenFromCookie(req.getHeader('cookie'))
 
   const token = cookieToken || headerToken